27. Instantiate at Runtime
Using Instantiate to create a clone of the explosion particle fx.

Created an empty game object to use it as a folder for the instantiated objects.

Used Transform.parent to put the instantiated objects to the empty game object.

Created a script to self-destruct for an x amount of time.